% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_write_geo.R
\name{arrow_space_dataset_chunks}
\alias{arrow_space_dataset_chunks}
\title{arrow_space_dataset_chunks}
\usage{
arrow_space_dataset_chunks(grid_rast, grid_chunks = 1)
}
\arguments{
\item{grid_rast}{terra \code{SpatRaster}. Grid defining the study area.}

\item{grid_chunks}{integer.
The number of chunks to partition the grid into for parallel processing.}
}
\description{
arrow_space_dataset_chunks
}
