% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow_space_dataset_copy.R
\name{arrow_space_dataset_copy}
\alias{arrow_space_dataset_copy}
\title{arrow_space_dataset_copy}
\usage{
arrow_space_dataset_copy(
  dataset_dir = NULL,
  dataset_name,
  template_name = dataset_name,
  table_names = NULL,
  overwrite = FALSE,
  skip_missing = FALSE,
  dataset_path = file.path(dataset_dir, dataset_name),
  template_path = file.path(dataset_dir, template_name)
)
}
\arguments{
\item{dataset_dir}{character. Path to \code{arrow_space} dataset directory.}

\item{dataset_name}{character. name of \code{arrow_space} dataset directory.}

\item{template_name}{character. name of \code{arrow_space} dataset directory to copy table from.}

\item{table_names}{character. TODO}

\item{overwrite}{logical. TODO}

\item{skip_missing}{logical. TODO}

\item{dataset_path}{character. Path to \code{arrow_space} dataset.
Defaults to \code{file.path(dataset_dir, dataset_name)}}

\item{template_path}{character. Path to \code{arrow_space} dataset to copy table from.
Defaults to \code{file.path(dataset_dir, template_name)}}
}
\description{
arrow_space_dataset_copy
}
\keyword{internal}
