% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow_space_dataset_copy_geo.R
\name{arrow_space_dataset_copy_geo}
\alias{arrow_space_dataset_copy_geo}
\title{arrow_space_dataset_copy_geo}
\usage{
arrow_space_dataset_copy_geo(
  dataset_dir = NULL,
  dataset_name,
  template_name,
  copy_raster_index_data = FALSE,
  partitions = NULL,
  tags = NULL,
  dataset_path = file.path(dataset_dir, dataset_name),
  template_path = file.path(dataset_dir, template_name),
  ...
)
}
\arguments{
\item{dataset_dir}{character. Path to \code{arrow_space} dataset directory.}

\item{dataset_name}{character. name of \code{arrow_space} dataset directory.}

\item{template_name}{character. name of \code{arrow_space} dataset directory to copy geo metadata from.}

\item{copy_raster_index_data}{logical. TODO}

\item{partitions}{character. TODO}

\item{tags}{character. TODO}

\item{dataset_path}{character. Path to \code{arrow_space} dataset.
Defaults to \code{file.path(dataset_dir, dataset_name)}}

\item{template_path}{character. Path to \code{arrow_space} dataset to copy geo metadata from.
Defaults to \code{file.path(dataset_dir, template_name)}}

\item{...}{unused}
}
\description{
arrow_space_dataset_copy_geo
}
\keyword{internal}
