% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow_space_dataset_write_geo.R
\name{arrow_space_dataset_write_geo}
\alias{arrow_space_dataset_write_geo}
\title{arrow_space_dataset_write_geo}
\usage{
arrow_space_dataset_write_geo(
  dataset_dir = NULL,
  dataset_name,
  chunks,
  geo_metadata,
  partitions = NULL,
  tags = NULL,
  dataset_path = file.path(dataset_dir, dataset_name)
)
}
\arguments{
\item{dataset_dir}{character. Path to \code{arrow_space} dataset directory.}

\item{dataset_name}{character. name of \code{arrow_space} dataset directory.}

\item{chunks}{data.table.}

\item{geo_metadata}{data.table.}

\item{partitions}{character. TODO}

\item{tags}{character. TODO}

\item{dataset_path}{character. Path to \code{arrow_space} dataset.
Defaults to \code{file.path(dataset_dir, dataset_name)}}
}
\description{
arrow_space_dataset_write_geo
}
