% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_data_copy.R
\name{cbm4_data_copy}
\alias{cbm4_data_copy}
\title{CBM4 data copy}
\usage{
cbm4_data_copy(
  template_data,
  cbm4_data,
  dataset_names = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{template_data}{character.
Path to template CBM4 spatial parquet datasets directory.
May be omitted if full paths to template datasets are provided.}

\item{cbm4_data}{character.
Path to CBM4 spatial parquet datasets directory.
May be omitted if full paths to datasets are provided.}

\item{dataset_names}{character. Names of CBM4 spatial parquet datasets to copy.}

\item{overwrite}{logical. Overwrite existing files.}
}
\value{
\code{NULL}. Data directory will be copied.
}
\description{
Copy a CBM4 spatial parquet datasets directory.
}
