% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_data_copy.R
\name{cbm4_data_copy_dataset}
\alias{cbm4_data_copy_dataset}
\title{CBM4 data copy dataset}
\usage{
cbm4_data_copy_dataset(
  cbm4_data = NULL,
  dataset_name = NULL,
  dataset_path = file.path(cbm4_data, dataset_name),
  template_data = cbm4_data,
  template_name = dataset_name,
  template_path = file.path(template_data, template_name),
  ...
)
}
\arguments{
\item{cbm4_data}{character.
Path to CBM4 spatial parquet datasets directory.
May be omitted if full paths to datasets are provided.}

\item{dataset_name}{character.
Name of the CBM4 spatial parquet dataset.}

\item{dataset_path}{character.
Path to the CBM4 spatial parquet dataset.
Defaults to \code{file.path(cbm4_data, dataset_name)}}

\item{template_data}{character.
Path to template CBM4 spatial parquet datasets directory.
May be omitted if full paths to template datasets are provided.}

\item{template_name}{character.
Name of a CBM4 spatial parquet dataset to use as a template for the new dataset.}

\item{template_path}{character.
Path to CBM4 spatial parquet dataset to use as a template for the new dataset.
Defaults to \code{file.path(cbm4_data, template_name)}}

\item{...}{arguments to \code{\link{arrow_space_dataset_copy}}}
}
\value{
\code{NULL}. Dataset will be copied.
}
\description{
Copy a CBM4 spatial parquet dataset.
}
