% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_write_disturbance.R
\name{cbm4_format_disturbance}
\alias{cbm4_format_disturbance}
\title{CBM4 format disturbance}
\usage{
cbm4_format_disturbance(
  grid_meta,
  dist_events,
  dist_meta = NULL,
  classifiers = NULL,
  def_enable_merge = TRUE,
  def_proportion = 1L,
  cbm_defaults_db = getOption("CBM4r.db.path"),
  ...
)
}
\arguments{
\item{grid_meta}{data.table. Grid metadata.
This table can be created with \code{cbm4_grid_meta} or \code{cbm4_set_grid_meta}.}

\item{dist_events}{data.table. Disturbance events.}

\item{dist_meta}{data.table. Disturbance metadata.
Required if \code{dist_events} does not contain a \code{disturbance_type} or \code{disturbance_type_id}.
If provided, \code{dist_events} and \code{dist_meta} must be linked by a \code{disturbance_id} column.}

\item{classifiers}{character.
Column names of cohort inventory identifiers.}

\item{def_enable_merge}{logical. Enable disturbance events to merge cohorts sharing the same new state.}

\item{def_proportion}{integer. Proportion of the pixel to be affected by the disturbance event.}

\item{cbm_defaults_db}{character.
Path to CBM defaults SQLite database.}

\item{...}{unused}
}
\value{
list with items:
\strong{index}: \code{arrow_space} raster indexed \code{data.table};
\strong{flat}: \code{arrow_space} flattened dataset \code{data.table}
}
\description{
CBM4 format disturbance
}
