% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_write_geo.R
\name{cbm4_format_grid_meta}
\alias{cbm4_format_grid_meta}
\title{cbm4_format_grid_meta}
\usage{
cbm4_format_grid_meta(
  grid_meta,
  cbm_defaults_db = NULL,
  def_afforestation_pre_type = "None",
  def_historic_disturbance_type = "Wildfire",
  def_last_pass_disturbance_type = "Wildfire",
  ...
)
}
\arguments{
\item{grid_meta}{data.table. Grid metadata.
Required columns: \code{pixel_index},
\code{admin_boundary} or \code{admin_boundary_id}.
\code{eco_boundary} or \code{eco_boundary_id},
\code{spatial_unit}.
Optional columns: \code{chunk_index}, \code{raster_index},
\code{afforestation_pre_type}, \code{historic_disturbance_type}, \code{last_pass_disturbance_type}.}

\item{cbm_defaults_db}{character.
Path to CBM defaults SQLite database.}

\item{def_afforestation_pre_type}{character. Land use before forestation.
Defined in CBM defaults database tables 'afforestation_pre_type'}

\item{def_historic_disturbance_type}{character. Historic disturbance type.
Defined in CBM defaults database tables 'disturbance_type' and 'disturbance_type_tr'.}

\item{def_last_pass_disturbance_type}{character. Last pass disturbance.
Defined in CBM defaults database tables 'disturbance_type' and 'disturbance_type_tr'.}

\item{...}{unused}
}
\description{
cbm4_format_grid_meta
}
