% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_write_parameters.R
\name{cbm4_format_increments}
\alias{cbm4_format_increments}
\title{CBM4 format increments}
\usage{
cbm4_format_increments(
  gc_incr,
  gc_meta = NULL,
  classifiers = NULL,
  long = TRUE,
  cbm_defaults_db = getOption("CBM4r.db.path")
)
}
\arguments{
\item{gc_incr}{data.table. Growth curve carbon increments.}

\item{gc_meta}{data.table. Growth curve metadata.
Required if \code{gc_incr} does not contain a \code{sw} flag or other classifiers.
If provided, \code{gc_events} and \code{gc_meta} must be linked by a \code{gc_id} column.}

\item{classifiers}{character.
Column names of cohort inventory identifiers.}

\item{long}{logical. Format table long or wide.}

\item{cbm_defaults_db}{character.
Path to CBM defaults SQLite database.}
}
\value{
data.table
}
\description{
CBM4 format increments
}
\keyword{internal}
