% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_write_inventory.R
\name{cbm4_format_inventory}
\alias{cbm4_format_inventory}
\title{CBM4 format inventory}
\usage{
cbm4_format_inventory(
  cohortDT,
  pixelDT,
  def_delay = 0L,
  def_land_class = "UNFCCC_FL_R_FL",
  def_cohort_proportion = 1L,
  area_unit_conversion = 1e-04,
  ...
)
}
\arguments{
\item{cohortDT}{data.table. Cohort inventory.}

\item{pixelDT}{character.
Table of pixel attributes.}

\item{def_delay}{integer. Regeneration delay.}

\item{def_land_class}{character. Land class code.
Defined in CBM defaults database tables 'land_class' and 'land_class_tr'.}

\item{def_cohort_proportion}{integer. A value between 0-1.
Percentage of the pixel's area that is attributed to the cohort.}

\item{area_unit_conversion}{numeric. Conversion factor of area to hectares (ha).}

\item{...}{unused}
}
\value{
list with items:
\strong{index}: \code{arrow_space} raster indexed \code{data.table};
\strong{flat}: \code{arrow_space} flattened dataset \code{data.table}
}
\description{
CBM4 format inventory
}
