% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_step_with_cohorts.R
\name{cbm4_format_simulation_cohorts}
\alias{cbm4_format_simulation_cohorts}
\title{CBM4 format simulation inventory}
\usage{
cbm4_format_simulation_cohorts(
  cohorts,
  timestep,
  classifiers,
  grid_meta = NULL,
  def_cohort_proportion = 1L,
  def_enabled = 1L,
  def_growth_enabled = 1L,
  def_growth_multiplier = 1L,
  def_regeneration_delay = 0L,
  def_land_class = "UNFCCC_FL_R_FL",
  area_unit_conversion = 0.0001
)
}
\arguments{
\item{cohorts}{data.table. Cohort inventory.}

\item{timestep}{integer. Simulation timestep with 1 representing the first year.}

\item{classifiers}{character.
Column names of cohort inventory identifiers.}

\item{grid_meta}{data.table. Grid metadata.
May not be required but can be provided for efficiency.
This table can be created with \code{cbm4_grid_meta} or \code{cbm4_set_grid_meta}.}

\item{def_cohort_proportion}{integer. A value between 0-1.
Percentage of the pixel's area that is attributed to the cohort.}

\item{def_enabled}{integer. TODO}

\item{def_growth_enabled}{integer. TODO}

\item{def_growth_multiplier}{integer. TODO}

\item{def_regeneration_delay}{integer. Regeneration delay.}

\item{def_land_class}{character. Land class code.
Defined in CBM defaults database tables 'land_class' and 'land_class_tr'.}

\item{area_unit_conversion}{numeric. Conversion factor of area to hectares (ha).}
}
\value{
list with items:
\strong{index}: \code{arrow_space} raster indexed \code{data.table};
\strong{flat}: \code{arrow_space} flattened dataset \code{data.table}
}
\description{
CBM4 format simulation inventory
}
\keyword{internal}
