% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_write_simulation_inventory.R
\name{cbm4_format_simulation_inventory}
\alias{cbm4_format_simulation_inventory}
\title{CBM4 format simulation inventory}
\usage{
cbm4_format_simulation_inventory(
  cohortDT,
  grid_meta,
  timestep,
  classifiers = NULL,
  def_cohort_proportion = 1L,
  area_unit_conversion = 1e-04,
  col_ignore = NULL,
  ...
)
}
\arguments{
\item{cohortDT}{data.table. Cohort inventory.}

\item{grid_meta}{data.table. Grid metadata.}

\item{timestep}{integer. Simulation timestep with 1 representing the first year.}

\item{classifiers}{character.
Column names of inventory identifiers.}

\item{def_cohort_proportion}{integer. A value between 0-1.
Percentage of the pixel's area that is attributed to the cohort.}

\item{area_unit_conversion}{numeric. Conversion factor of area to hectares (ha).}

\item{col_ignore}{character. Names of \code{cohortDT} columns to exclude}

\item{...}{unused}
}
\value{
list with items:
\strong{index}: \code{arrow_space} raster indexed \code{data.table};
\strong{flat}: \code{arrow_space} flattened dataset \code{data.table}
}
\description{
CBM4 format simulation inventory
}
