% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_read_geo.R
\name{cbm4_read_geo}
\alias{cbm4_read_geo}
\title{CBM read geo}
\usage{
cbm4_read_geo(
  cbm4_results,
  dataset_name = "simulation",
  dataset_path = file.path(cbm4_results, dataset_name)
)
}
\arguments{
\item{cbm4_results}{character or \code{SQLResultsProcessor}.
Path to CBM4 spatial datasets directory
or a \code{SQLResultsProcessor} object created with \code{\link{cbm4_results_processor}}.}

\item{dataset_name}{character.
Name of the CBM4 spatial parquet dataset.}

\item{dataset_path}{character.
Path to the CBM4 spatial parquet dataset.
Defaults to \code{file.path(cbm4_data, dataset_name)}}
}
\value{
\code{SpatRaster}
}
\description{
Recreate a study area grid from a CBM4 spatial parquet dataset.
}
