% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_results_by_pixel.R
\name{cbm4_results_by_pixel}
\alias{cbm4_results_by_pixel}
\title{CBM4 results: by pixel}
\usage{
cbm4_results_by_pixel(cbm4_results, queryName, timestep = NULL)
}
\arguments{
\item{cbm4_results}{character or \code{SQLResultsProcessor}.
Path to CBM4 spatial datasets directory
or a \code{SQLResultsProcessor} object created with \code{\link{cbm4_results_processor}}.}

\item{queryName}{character. Preset SQL query file name}

\item{timestep}{integer. Simulation timestep with 1 representing the first year.}
}
\value{
\code{data.table}
}
\description{
CBM4 results: by pixel
}
\keyword{internal}
