% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_results_by_timestep.R
\name{cbm4_results_emissions_by_timestep}
\alias{cbm4_results_emissions_by_timestep}
\title{CBM4 results: Emissions by timestep}
\usage{
cbm4_results_emissions_by_timestep(cbm4_data, timestep = NULL)
}
\arguments{
\item{cbm4_data}{character.
Path to CBM4 spatial parquet datasets directory.
May be omitted if full paths to datasets are provided.}

\item{timestep}{integer. Simulation timestep with 1 representing the first year.}
}
\value{
\code{data.table}
}
\description{
CBM4 results: Emissions by timestep
}
