% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_results_grid.R
\name{cbm4_results_grid_key}
\alias{cbm4_results_grid_key}
\title{CBM4 results grid key}
\usage{
cbm4_results_grid_key(
  cbm4_results,
  dataset_name = "simulation",
  coords = FALSE
)
}
\arguments{
\item{cbm4_results}{character or \code{SQLResultsProcessor}.
Path to CBM4 spatial datasets directory
or a \code{SQLResultsProcessor} object created with \code{\link{cbm4_results_processor}}.}

\item{dataset_name}{character.
Name of the CBM4 spatial parquet dataset.}

\item{coords}{logical. Return pixel coordinates.}
}
\value{
\code{data.table} with columns \code{pixel_index}, \code{chunk_index}, \code{raster_index}.
}
\description{
Retrieve a study area grid key for a CBM4 spatial parquet dataset.
}
