% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_results_query.R
\name{cbm4_results_query}
\alias{cbm4_results_query}
\title{CBM4 results query}
\usage{
cbm4_results_query(cbm4_data, query)
}
\arguments{
\item{cbm4_data}{character.
Path to CBM4 spatial parquet datasets directory.
May be omitted if full paths to datasets are provided.}

\item{query}{character. SQL query}
}
\description{
Query CBM4 results with SQL.
}
