% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_results_query.R
\name{cbm4_results_query_preset}
\alias{cbm4_results_query_preset}
\title{CBM4 results query preset}
\usage{
cbm4_results_query_preset(cbm4_data, queryName, where = NULL)
}
\arguments{
\item{cbm4_data}{character.
Path to CBM4 spatial parquet datasets directory.
May be omitted if full paths to datasets are provided.}

\item{queryName}{character. Preset SQL query file name}
}
\description{
Query results from CBM4r preset SQL query file.
}
\keyword{internal}
