% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_results_raster.R
\name{cbm4_results_raster}
\alias{cbm4_results_raster}
\title{CBM4 results raster}
\usage{
cbm4_results_raster(
  cbm4_results,
  view_name = NULL,
  view_column = NULL,
  timesteps = NULL,
  list = FALSE
)
}
\arguments{
\item{cbm4_results}{character or \code{SQLResultsProcessor}.
Path to CBM4 spatial datasets directory
or a \code{SQLResultsProcessor} object created with \code{\link{cbm4_results_processor}}.}

\item{view_name}{character. \code{SQLResultsProcessor} view name.
If NULL the function will return an empty study area grid.}

\item{view_column}{character. \code{SQLResultsProcessor} view column name.}

\item{timesteps}{integer. Simulation timesteps with 1 representing the first year.}

\item{list}{logical. Return a table of options.}
}
\value{
\code{SpatRaster}
}
\description{
Read simulation results into a raster with carbon in tonnes per hectare (t/ha).
}
