% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_results_totals.R
\name{cbm4_results_totals}
\alias{cbm4_results_totals}
\title{CBM4 results totals}
\usage{
cbm4_results_totals(
  cbm4_results,
  view_name,
  view_columns = NULL,
  timesteps = NULL,
  units = "t",
  list = FALSE
)
}
\arguments{
\item{cbm4_results}{character or \code{SQLResultsProcessor}.
Path to CBM4 spatial datasets directory
or a \code{SQLResultsProcessor} object created with \code{\link{cbm4_results_processor}}.}

\item{view_name}{character. \code{SQLResultsProcessor} view name.}

\item{view_columns}{character. \code{SQLResultsProcessor} view column names.
The vector can have names to use as column aliases.}

\item{timesteps}{integer. Simulation timesteps with 1 representing the first year.}

\item{units}{character. "t", "Mt", or "t/ha".}

\item{list}{logical. Return a table of options.}
}
\value{
\code{data.table}
}
\description{
Read simulation results into a table with carbon totals per timestep.
}
