% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_set_db_path.R
\name{cbm4_set_db_path}
\alias{cbm4_set_db_path}
\title{CBM4 set database path}
\usage{
cbm4_set_db_path(cbm_defaults_db)
}
\arguments{
\item{cbm_defaults_db}{character.
Path to CBM defaults SQLite database.}
}
\description{
Set custom parameters by using a modified version of the CBM defaults RSQLite database.
This path can instead be passed directly to functions that use it.
}
