% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_spinup.R
\name{cbm4_spinup}
\alias{cbm4_spinup}
\title{CBM4 spinup}
\usage{
cbm4_spinup(
  cbm4_data = NULL,
  max_workers = NULL,
  cbm_defaults_db = getOption("CBM4r.db.path"),
  spinup_parameters_dataset = file.path(cbm4_data, "spinup_parameters"),
  inventory_dataset = file.path(cbm4_data, "inventory"),
  simulation_dataset = file.path(cbm4_data, "simulation")
)
}
\arguments{
\item{cbm4_data}{character.
Path to CBM4 spatial parquet datasets directory.
May be omitted if full paths to datasets are provided.}

\item{max_workers}{integer.
Number of parallel processes to use.
NULL or NA will bypass this and and cap processing at the system's available resources.
A positive integer will spawn multiple processes assigned to each \code{chunk_index}.}

\item{cbm_defaults_db}{character.
Path to CBM defaults SQLite database.}

\item{spinup_parameters_dataset}{character.
Path to spinup parameters CBM4 spatial parquet dataset.
Defaults to \code{file.path(cbm4_data, "spinup_parameters")}}

\item{inventory_dataset}{character.
Path to inventory CBM4 spatial parquet dataset.
Defaults to \code{file.path(cbm4_data, "inventory")}}

\item{simulation_dataset}{character.
Path to simulation CBM4 spatial parquet dataset.
Defaults to \code{file.path(cbm4_data, "simulation")}}
}
\value{
\code{NULL}. Updates will be made to CBM4 spatial parquet datasets.
}
\description{
Run spinup initialization on CBM4 spatial parquet datasets.
}
