% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_step.R
\name{cbm4_step}
\alias{cbm4_step}
\title{CBM4 step}
\usage{
cbm4_step(
  cbm4_data = NULL,
  timestep,
  area_unit_conversion = 0.0001,
  area_validation = NULL,
  write_parameters = NULL,
  max_workers = NULL,
  cbm_defaults_db = getOption("CBM4r.db.path"),
  step_parameters_dataset = file.path(cbm4_data, "step_parameters"),
  inventory_dataset = file.path(cbm4_data, "inventory"),
  disturbance_dataset = file.path(cbm4_data, "disturbance"),
  simulation_dataset = file.path(cbm4_data, "simulation"),
  ...
)
}
\arguments{
\item{cbm4_data}{character.
Path to CBM4 spatial parquet datasets directory.
May be omitted if full paths to datasets are provided.}

\item{timestep}{integer. Simulation timestep with 1 representing the first year.}

\item{area_unit_conversion}{numeric. Conversion factor of area to hectares (ha).}

\item{area_validation}{logical. Validate proportional area totals.}

\item{write_parameters}{logical. Write cohort step parameters to file.}

\item{max_workers}{integer.
Number of parallel processes to use.
NULL or NA will bypass this and and cap processing at the system's available resources.
A positive integer will spawn multiple processes assigned to each \code{chunk_index}.}

\item{cbm_defaults_db}{character.
Path to CBM defaults SQLite database.}

\item{step_parameters_dataset}{character.
Path to step parameters CBM4 spatial parquet dataset.
Defaults to \code{file.path(cbm4_data, "step_parameters")}}

\item{inventory_dataset}{character.
Path to inventory CBM4 spatial parquet dataset.
Defaults to \code{file.path(cbm4_data, "inventory")}}

\item{disturbance_dataset}{character.
Path to disturbance CBM4 spatial parquet dataset.
Defaults to \code{file.path(cbm4_data, "disturbance")}}

\item{simulation_dataset}{character.
Path to simulation CBM4 spatial parquet dataset.
Defaults to \code{file.path(cbm4_data, "simulation")}}

\item{...}{unused}
}
\value{
\code{NULL}. Updates will be made to CBM4 spatial parquet datasets.
}
\description{
Run an annual step on CBM4 spatial parquet datasets.
}
