% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_step_with_cohorts.R
\name{cbm4_step_with_cohorts}
\alias{cbm4_step_with_cohorts}
\title{CBM4 step: with cohorts}
\usage{
cbm4_step_with_cohorts(
  cbm4_data = NULL,
  timestep,
  simulation_dataset = file.path(cbm4_data, "simulation"),
  ...
)
}
\arguments{
\item{cbm4_data}{character.
Path to CBM4 spatial parquet datasets directory.
May be omitted if full paths to datasets are provided.}

\item{timestep}{integer. Simulation timestep with 1 representing the first year.}

\item{simulation_dataset}{character.
Path to simulation CBM4 spatial parquet dataset.
Defaults to \code{file.path(cbm4_data, "simulation")}}

\item{...}{arguments to \code{\link{cbm4_write_simulation}} or \code{\link{cbm4_step}}}
}
\value{
\code{NULL}. Updates will be made to CBM4 spatial parquet datasets.
}
\description{
Run an annual step on CBM4 spatial parquet datasets with an alternate set of cohort data.
}
