% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_summary_products.R
\name{cbm4_summary_products}
\alias{cbm4_summary_products}
\title{CBM4 summary: products}
\usage{
cbm4_summary_products(
  cbm4_data = NULL,
  timestep = NULL,
  simulation_dataset = file.path(cbm4_data, "simulation")
)
}
\arguments{
\item{cbm4_data}{character.
Path to CBM4 spatial parquet datasets directory.
May be omitted if full paths to datasets are provided.}

\item{timestep}{integer. Simulation timestep with 1 representing the first year.}

\item{simulation_dataset}{character.
Path to simulation CBM4 spatial parquet dataset.
Defaults to \code{file.path(cbm4_data, "simulation")}}
}
\value{
\code{data.table}
}
\description{
CBM4 summary: products
}
