% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_virtualenv_create.R
\name{cbm4_virtualenv_create}
\alias{cbm4_virtualenv_create}
\title{cbm4_virtualenv_create}
\usage{
cbm4_virtualenv_create(
  envname,
  version = NULL,
  upgrade = FALSE,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{envname}{character. Virtual environment name.}

\item{version}{character. CBM4 version.}

\item{upgrade}{logical. Upgrade Python packages.}

\item{quiet}{logical. Silence pip output.}

\item{...}{arguments to \code{\link[reticulate]{virtualenv_create}}}
}
\description{
Create a Python virtual environment to run CBM4.
}
