% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_write_geo.R
\name{cbm4_write_geo}
\alias{cbm4_write_geo}
\title{CBM write geo}
\usage{
cbm4_write_geo(
  cbm4_data = NULL,
  dataset_name,
  template_name = NULL,
  grid_rast = NULL,
  grid_chunks = 1,
  write_pixels = TRUE,
  dataset_path = file.path(cbm4_data, dataset_name),
  template_path = file.path(cbm4_data, template_name),
  ...
)
}
\arguments{
\item{cbm4_data}{character.
Path to CBM4 spatial parquet datasets directory.
May be omitted if full paths to datasets are provided.}

\item{dataset_name}{character.
Name of the CBM4 spatial parquet dataset.}

\item{template_name}{character.
Name of a CBM4 spatial parquet dataset to use as a template for the new dataset.}

\item{grid_rast}{terra \code{SpatRaster}. Grid defining the study area.}

\item{grid_chunks}{integer.
The number of chunks to partition the grid into for parallel processing.}

\item{write_pixels}{logical. Write pixel table to file.}

\item{dataset_path}{character.
Path to the CBM4 spatial parquet dataset.
Defaults to \code{file.path(cbm4_data, dataset_name)}}

\item{template_path}{character.
Path to CBM4 spatial parquet dataset to use as a template for the new dataset.
Defaults to \code{file.path(cbm4_data, template_name)}}

\item{...}{arguments to \code{\link{arrow_space_dataset_write_geo}}
or \code{\link{arrow_space_dataset_copy_geo}}}
}
\value{
\code{NULL}. Data will be written to the CBM4 spatial parquet dataset.
}
\description{
Initiate a CBM4 spatial parquet dataset with study area geographic metadata.
This can be done by providing a raster grid or the name of a dataset to use as a template.
The dataset will contain an additional "pixels" table
mapping grid pixel locations to their parallel processing partition chunks.
This table has columns \code{pixel_index}, \code{chunk_index}, \code{raster_index}, \code{x}, \code{y}, \code{area}.
}
