% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_write_simulation.R
\name{cbm4_write_simulation}
\alias{cbm4_write_simulation}
\title{CBM4 write simulation}
\usage{
cbm4_write_simulation(
  cbm4_data = NULL,
  timestep = 0,
  cohorts = NULL,
  classifiers = NULL,
  grid_meta = NULL,
  grid_rast = NULL,
  template_name = "inventory",
  template_path = file.path(cbm4_data, template_name),
  dataset_name = "simulation",
  dataset_path = file.path(cbm4_data, dataset_name),
  cbm_defaults_db = getOption("CBM4r.db.path"),
  schema = NULL,
  ...
)
}
\arguments{
\item{cbm4_data}{character.
Path to CBM4 spatial parquet datasets directory.
May be omitted if full paths to datasets are provided.}

\item{timestep}{integer. Simulation timestep with 1 representing the first year.}

\item{cohorts}{data.table. Cohort inventory.}

\item{classifiers}{character.
Column names of cohort inventory identifiers.}

\item{grid_meta}{data.table. Grid metadata.
This table can be created with \code{cbm4_grid_meta} or \code{cbm4_set_grid_meta}.}

\item{grid_rast}{terra \code{SpatRaster}. Grid defining the study area.}

\item{template_name}{character.
Name of a CBM4 spatial parquet dataset to use as a template for the new dataset.}

\item{template_path}{character.
Path to CBM4 spatial parquet dataset to use as a template for the new dataset.
Defaults to \code{file.path(cbm4_data, template_name)}}

\item{dataset_name}{character.
Name of the CBM4 spatial parquet dataset.}

\item{dataset_path}{character.
Path to the CBM4 spatial parquet dataset.
Defaults to \code{file.path(cbm4_data, dataset_name)}}

\item{cbm_defaults_db}{character.
Path to CBM defaults SQLite database.}

\item{schema}{internal}

\item{...}{arguments to \code{\link{cbm4_format_simulation}}}
}
\value{
\code{NULL}. Data will be written to the CBM4 spatial parquet dataset.
}
\description{
Write inventory and associated carbon pools
directly to a simulation CBM4 spatial parquet dataset.
}
