% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_step_with_cohorts.R
\name{cbm4_write_simulation_cohorts}
\alias{cbm4_write_simulation_cohorts}
\title{CBM4 write simulation cohorts}
\usage{
cbm4_write_simulation_cohorts(
  cbm4_data = NULL,
  cohorts,
  grid_meta = NULL,
  timestep,
  dataset_name = "simulation",
  dataset_path = file.path(cbm4_data, dataset_name),
  schema = NULL,
  ...
)
}
\arguments{
\item{cbm4_data}{character.
Path to CBM4 spatial parquet datasets directory.
May be omitted if full paths to datasets are provided.}

\item{cohorts}{data.table. Cohort inventory.}

\item{grid_meta}{data.table. Grid metadata.
May not be required but can be provided for efficiency.
This table can be created with \code{cbm4_grid_meta} or \code{cbm4_set_grid_meta}.}

\item{timestep}{integer. Simulation timestep with 1 representing the first year.}

\item{dataset_name}{character.
Name of the CBM4 spatial parquet dataset.}

\item{dataset_path}{character.
Path to the CBM4 spatial parquet dataset.
Defaults to \code{file.path(cbm4_data, dataset_name)}}

\item{...}{arguments to \code{\link{cbm4_format_simulation_cohorts}}}
}
\value{
\code{NULL}. Data will be written to the CBM4 spatial parquet dataset.
}
\description{
Write cohort inventory to a simulation CBM4 spatial parquet dataset.
}
\keyword{internal}
