% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cbm_defaults_db.R
\name{cbm_defaults_listTables}
\alias{cbm_defaults_listTables}
\title{CBM defaults: list tables}
\usage{
cbm_defaults_listTables(cbm_defaults_db = getOption("CBM4r.db.path"), ...)
}
\arguments{
\item{cbm_defaults_db}{character.
Path to CBM defaults SQLite database.}

\item{...}{arguments to \code{RSQLite::dbListTables}}
}
\description{
List tables in a CBM defaults SQLite database.
}
