% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cbm_defaults_db.R
\name{cbm_defaults_readTable}
\alias{cbm_defaults_readTable}
\title{CBM defaults: read table}
\usage{
cbm_defaults_readTable(
  tableName,
  cbm_defaults_db = getOption("CBM4r.db.path"),
  localeID = getOption("CBM4r.db.localeID"),
  ...
)
}
\arguments{
\item{tableName}{character. Table name.
Use \code{cbm_defaults_listTables} to select a table.}

\item{cbm_defaults_db}{character.
Path to CBM defaults SQLite database.}

\item{localeID}{integer. Locale ID.}

\item{...}{arguments to \code{RSQLite::dbReadTable}}
}
\description{
Read a table from a CBM defaults SQLite database.
}
