% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_results_by_timestep.R
\name{.cbm4_results_by_timestep}
\alias{.cbm4_results_by_timestep}
\title{CBM4 results: by timestep}
\usage{
.cbm4_results_by_timestep(
  cbm4_results,
  type,
  units,
  timesteps = NULL,
  alias = NULL
)
}
\arguments{
\item{cbm4_results}{character or \code{SQLResultsProcessor}.
Path to CBM4 spatial datasets directory
or a \code{SQLResultsProcessor} object created with \code{\link{cbm4_results_processor}}.}

\item{type}{character. Results query type. e.g. 'pools'}

\item{units}{character. "t/ha", "t", or "Mt".}

\item{timesteps}{integer. Simulation timesteps with 1 representing the first year.}
}
\value{
\code{data.table}
}
\description{
CBM4 results: by timestep
}
\keyword{internal}
