% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_write_disturbance.R
\name{cbm4_format_disturbance}
\alias{cbm4_format_disturbance}
\title{CBM4 format disturbance}
\usage{
cbm4_format_disturbance(
  distMeta,
  distEvents,
  grid_meta,
  classifiers = NULL,
  def_proportion = 1L,
  def_enable_merge = 0L,
  def_sort_id = 0L,
  def_filter_id = 0L,
  def_undisturbed_transition_id = 0L,
  def_disturbed_transition_id = 0L,
  cbm_defaults_db = getOption("CBM4r.db.path"),
  ...
)
}
\arguments{
\item{distMeta}{data.table. Disturbance metadata.}

\item{distEvents}{data.table. Disturbance events.}

\item{grid_meta}{data.table. Grid metadata.}

\item{classifiers}{character.
Column names of inventory identifiers.}

\item{def_proportion}{integer. TODO}

\item{def_enable_merge}{integer. TODO}

\item{def_sort_id}{integer. TODO}

\item{def_filter_id}{integer. TODO}

\item{def_undisturbed_transition_id}{integer. Set to 0 to indicate no transitions.}

\item{def_disturbed_transition_id}{integer. Set to 0 to indicate no transitions.}

\item{cbm_defaults_db}{character.
Path to CBM defaults SQLite database.}

\item{...}{unused}
}
\value{
list with items:
\strong{index}: \code{arrow_space} raster indexed \code{data.table};
\strong{flat}: \code{arrow_space} flattened dataset \code{data.table}
}
\description{
CBM4 format disturbance
}
