% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_write_parameters.R
\name{cbm4_format_increments}
\alias{cbm4_format_increments}
\title{CBM4 format increments}
\usage{
cbm4_format_increments(
  classifiers,
  gcMeta,
  gcIncr,
  long = TRUE,
  cbm_defaults_db = NULL
)
}
\arguments{
\item{classifiers}{character.
Column names of inventory identifiers.}

\item{gcMeta}{data.table. Growth curve metadata}

\item{gcIncr}{data.table. Growth curve carbon increments}

\item{long}{logical. Format table long or wide.}

\item{cbm_defaults_db}{character.
Path to CBM defaults SQLite database.}
}
\value{
data.table
}
\description{
CBM4 format increments
}
