% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_write_inventory.R
\name{cbm4_format_inventory}
\alias{cbm4_format_inventory}
\title{CBM4 format inventory}
\usage{
cbm4_format_inventory(
  inventoryDT,
  pixelDT,
  cbm_defaults_db = NULL,
  def_land_class = "UNFCCC_FL_R_FL",
  def_afforestation_pre_type = "None",
  def_historic_disturbance_type = "Wildfire",
  def_last_pass_disturbance_type = "Wildfire",
  def_delay = 0L,
  def_cohort_proportion = 1L
)
}
\arguments{
\item{inventoryDT}{data.table. Cohort inventory.}

\item{pixelDT}{TODO}

\item{cbm_defaults_db}{character.
Path to CBM defaults SQLite database.}

\item{def_land_class}{character. Land class code.
Defined in CBM defaults database tables 'land_class' and 'land_class_tr'.}

\item{def_afforestation_pre_type}{character. Land use before forestation.
Defined in CBM defaults database tables 'afforestation_pre_type'}

\item{def_historic_disturbance_type}{character. Historic disturbance type.
Defined in CBM defaults database tables 'disturbance_type' and 'disturbance_type_tr'.}

\item{def_last_pass_disturbance_type}{character. Last pass disturbance.
Defined in CBM defaults database tables 'disturbance_type' and 'disturbance_type_tr'.}

\item{def_delay}{integer. Regeneration delay.}

\item{def_cohort_proportion}{integer. A value between 0-1.
Percentage of the pixel's area that is attributed to the cohort.}
}
\value{
list with items:
\strong{index}: \code{arrow_space} raster indexed \code{data.table};
\strong{flat}: \code{arrow_space} flattened dataset \code{data.table}
}
\description{
CBM4 format inventory
}
