% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_write_simulation_inventory.R
\name{cbm4_format_simulation_inventory}
\alias{cbm4_format_simulation_inventory}
\title{CBM4 format simulation inventory}
\usage{
cbm4_format_simulation_inventory(
  cohortDT,
  grid_meta,
  timestep,
  classifiers = NULL,
  ...
)
}
\arguments{
\item{cohortDT}{data.table. Cohort inventory.}

\item{grid_meta}{data.table. Grid metadata.}

\item{timestep}{integer. Simulation timestep with 1 representing the first year.}

\item{classifiers}{character.
Column names of inventory identifiers.}

\item{...}{arguments to \code{\link{cbm4_format_inventory}}}
}
\value{
list with items:
\strong{index}: \code{arrow_space} raster indexed \code{data.table};
\strong{flat}: \code{arrow_space} flattened dataset \code{data.table}
}
\description{
CBM4 format simulation inventory
}
