% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_write_simulation_inventory.R
\name{cbm4_format_simulation_inventory}
\alias{cbm4_format_simulation_inventory}
\title{CBM4 format simulation inventory}
\usage{
cbm4_format_simulation_inventory(
  cohorts,
  timestep,
  classifiers = NULL,
  grid_meta = NULL,
  col_ignore = NULL,
  area_unit_conversion = 1e-04,
  def_cohort_proportion = 1L,
  def_state.enabled = 1L,
  def_state.growth_enabled = 1L,
  def_state.growth_multiplier = 1L,
  def_state.regeneration_delay = 0L,
  def_state.land_class = "UNFCCC_FL_R_FL",
  ...
)
}
\arguments{
\item{cohorts}{data.table. Cohort inventory.}

\item{timestep}{integer. Simulation timestep with 1 representing the first year.}

\item{classifiers}{character.
Column names of cohort inventory identifiers.}

\item{grid_meta}{data.table. Grid metadata.
May not be required but can be provided for efficiency.
This table can be created with \code{cbm4_grid_meta} or \code{cbm4_set_grid_meta}.}

\item{col_ignore}{character. Names of \code{cohorts} columns to exclude}

\item{area_unit_conversion}{numeric. Conversion factor of area to hectares (ha).}

\item{def_cohort_proportion}{integer. A value between 0-1.
Percentage of the pixel's area that is attributed to the cohort.}

\item{def_state.enabled}{integer. TODO}

\item{def_state.growth_enabled}{integer. TODO}

\item{def_state.growth_multiplier}{integer. TODO}

\item{def_state.regeneration_delay}{integer. Regeneration delay.}

\item{def_state.land_class}{character. Land class code.
Defined in CBM defaults database tables 'land_class' and 'land_class_tr'.}

\item{...}{unused}
}
\value{
list with items:
\strong{index}: \code{arrow_space} raster indexed \code{data.table};
\strong{flat}: \code{arrow_space} flattened dataset \code{data.table}
}
\description{
CBM4 format simulation inventory
}
