% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_grid_meta.R
\name{cbm4_grid_meta}
\alias{cbm4_grid_meta}
\title{CBM4 grid metadata}
\usage{
cbm4_grid_meta(
  grid_rast,
  admin_boundary = NULL,
  eco_boundary = NULL,
  eco_boundary_id = NULL,
  def_afforestation_pre_type = "None",
  def_historic_disturbance_type = "Wildfire",
  def_last_pass_disturbance_type = "Wildfire",
  cbm_defaults_db = getOption("CBM4r.db.path"),
  ...
)
}
\arguments{
\item{grid_rast}{terra \code{SpatRaster}. Grid defining the study area.}

\item{admin_boundary}{character. Canada province or territory name.}

\item{eco_boundary}{character. Canada ecozone name.}

\item{eco_boundary_id}{integer. Canada ecozone ID. Provide this or \code{eco_boundary}.}

\item{def_afforestation_pre_type}{character. Land use before forestation.
Defined in CBM defaults database tables 'afforestation_pre_type'}

\item{def_historic_disturbance_type}{character. Historic disturbance type.
Defined in CBM defaults database tables 'disturbance_type' and 'disturbance_type_tr'.}

\item{def_last_pass_disturbance_type}{character. Last pass disturbance.
Defined in CBM defaults database tables 'disturbance_type' and 'disturbance_type_tr'.}

\item{cbm_defaults_db}{character.
Path to CBM defaults SQLite database.}

\item{...}{unused}
}
\description{
Create a study area grid metadata table.
}
