% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_grid_meta.R
\name{cbm4_grid_meta}
\alias{cbm4_grid_meta}
\title{CBM4 grid metadata}
\usage{
cbm4_grid_meta(
  grid_rast,
  admin_boundary = NULL,
  eco_boundary = NULL,
  eco_boundary_id = NULL,
  ...
)
}
\arguments{
\item{grid_rast}{terra \code{SpatRaster}. Grid defining the study area.}

\item{admin_boundary}{character. TODO}

\item{eco_boundary}{character. TODO}

\item{eco_boundary_id}{integer. TODO}

\item{...}{arguments to \code{\link{cbm4_set_grid_meta}}}
}
\description{
Create a study area grid metadata table.
}
