% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_read_simulation_inventory.R
\name{cbm4_read_simulation_inventory}
\alias{cbm4_read_simulation_inventory}
\title{CBM4 read simulation inventory}
\usage{
cbm4_read_simulation_inventory(cbm4_results, timestep)
}
\arguments{
\item{cbm4_results}{character or \code{SQLResultsProcessor}.
Path to CBM4 spatial datasets directory
or a \code{SQLResultsProcessor} object created with \code{\link{cbm4_results_processor}}.}

\item{timestep}{integer. Simulation timestep with 1 representing the first year.}
}
\value{
\code{data.table}
}
\description{
Read inventory from a simulation CBM4 spatial parquet dataset.
}
