% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_results_processor.R
\name{cbm4_results_processor}
\alias{cbm4_results_processor}
\title{CBM4 results processor}
\usage{
cbm4_results_processor(cbm4_data, views = TRUE)
}
\arguments{
\item{cbm4_data}{character.
Path to CBM4 spatial parquet datasets directory.
May be omitted if full paths to datasets are provided.}

\item{views}{logical. Require views to be available}
}
\value{
\code{SQLResultsProcessor}
}
\description{
CBM4 results processor
}
