% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_results_query.R
\name{cbm4_results_query}
\alias{cbm4_results_query}
\title{CBM4 results query}
\usage{
cbm4_results_query(cbm4_results, query = NULL)
}
\arguments{
\item{cbm4_results}{character or \code{SQLResultsProcessor}.
Path to CBM4 spatial datasets directory
or a \code{SQLResultsProcessor} object created with \code{\link{cbm4_results_processor}}.}

\item{query}{character. SQL query}
}
\value{
\code{data.table}
}
\description{
Query CBM4 results with SQL.
}
