% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_step_with_cohorts.R
\name{cbm4_step_with_cohorts}
\alias{cbm4_step_with_cohorts}
\title{CBM4 step: with cohorts}
\usage{
cbm4_step_with_cohorts(
  cbm4_data = NULL,
  timestep,
  cohorts = NULL,
  grid_meta = NULL,
  area_unit_conversion = 0.0001,
  def_cohort_proportion = 1L,
  def_enabled = 1L,
  def_growth_enabled = 1L,
  def_growth_multiplier = 1L,
  def_regeneration_delay = 0L,
  def_land_class = "UNFCCC_FL_R_FL",
  simulation_dataset = file.path(cbm4_data, "simulation"),
  ...
)
}
\arguments{
\item{cbm4_data}{character.
Path to CBM4 spatial parquet datasets directory.
May be omitted if full paths to datasets are provided.}

\item{timestep}{integer. Simulation timestep with 1 representing the first year.}

\item{cohorts}{data.table. Cohort inventory.}

\item{grid_meta}{data.table. Grid metadata.
May not be required but can be provided for efficiency.
This table can be created with \code{cbm4_grid_meta} or \code{cbm4_set_grid_meta}.}

\item{area_unit_conversion}{numeric. Conversion factor of area to hectares (ha).}

\item{def_cohort_proportion}{integer. A value between 0-1.
Percentage of the pixel's area that is attributed to the cohort.}

\item{def_enabled}{integer. TODO}

\item{def_growth_enabled}{integer. TODO}

\item{def_growth_multiplier}{integer. TODO}

\item{def_regeneration_delay}{integer. Regeneration delay.}

\item{def_land_class}{character. Land class code.
Defined in CBM defaults database tables 'land_class' and 'land_class_tr'.}

\item{simulation_dataset}{character.
Path to simulation CBM4 spatial parquet dataset.
Defaults to \code{file.path(cbm4_data, "simulation")}}

\item{...}{arguments to \code{\link{cbm4_step}}}
}
\value{
\code{NULL}. Updates will be made to CBM4 spatial parquet datasets.
}
\description{
Run an annual step on CBM4 spatial parquet datasets with an alternate set of cohort data.
}
