% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_virtualenv_create.R
\name{cbm4_virtualenv_create}
\alias{cbm4_virtualenv_create}
\title{CBM4 create virtualenv}
\usage{
cbm4_virtualenv_create(
  virtualenv,
  version = NULL,
  upgrade = FALSE,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{virtualenv}{character. TODO}

\item{version}{character. TODO}

\item{upgrade}{logical. TODO}

\item{quiet}{logical. TODO}

\item{...}{TODO}
}
\description{
Create a Python virtual environment to run CBM4.
}
