% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_write_disturbance.R
\name{cbm4_write_disturbance}
\alias{cbm4_write_disturbance}
\title{CBM4 write disturbance}
\usage{
cbm4_write_disturbance(
  cbm4_data = NULL,
  distMeta = NULL,
  distEvents = NULL,
  classifiers = NULL,
  grid_meta = NULL,
  template_name = "inventory",
  template_path = file.path(cbm4_data, template_name),
  dataset_name = "disturbance",
  dataset_path = file.path(cbm4_data, dataset_name),
  ...
)
}
\arguments{
\item{cbm4_data}{character.
Path to CBM4 spatial parquet datasets directory.
May be omitted if full paths to datasets are provided.}

\item{distMeta}{data.table. Disturbance metadata.}

\item{distEvents}{data.table. Disturbance events.}

\item{classifiers}{character.
Column names of inventory identifiers.}

\item{grid_meta}{data.table. Grid metadata.}

\item{template_name}{character.
Name of a CBM4 spatial parquet dataset to use as a template for the new dataset.}

\item{template_path}{character.
Path to CBM4 spatial parquet dataset to use as a template for the new dataset.
Defaults to \code{file.path(cbm4_data, template_name)}}

\item{dataset_name}{character.
Name of the CBM4 spatial parquet dataset.}

\item{dataset_path}{character.
Path to the CBM4 spatial parquet dataset.
Defaults to \code{file.path(cbm4_data, dataset_name)}}

\item{...}{arguments to \code{\link{cbm4_format_disturbance}}}
}
\value{
\code{NULL}. Data will be written to the CBM4 spatial parquet dataset.
}
\description{
Write disturbances to a CBM4 spatial parquet dataset.
}
