% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_write_geo.R
\name{cbm4_write_geo}
\alias{cbm4_write_geo}
\title{CBM4 write geo}
\usage{
cbm4_write_geo(
  cbm4_data = NULL,
  dataset_name,
  grid_meta,
  grid_rast,
  dataset_path = file.path(cbm4_data, dataset_name),
  ...
)
}
\arguments{
\item{cbm4_data}{character.
Path to CBM4 spatial parquet datasets directory.
May be omitted if full paths to datasets are provided.}

\item{dataset_name}{character.
Name of the CBM4 spatial parquet dataset.}

\item{grid_meta}{data.table. Grid metadata set with \code{cbm4_set_grid_meta}.}

\item{grid_rast}{terra \code{SpatRaster}. Grid defining the study area.}

\item{dataset_path}{character.
Path to the CBM4 spatial parquet dataset.
Defaults to \code{file.path(cbm4_data, dataset_name)}}

\item{...}{arguments to \code{\link{arrow_space_dataset_write_geo}}}
}
\value{
\code{NULL}. Data will be written to the CBM4 spatial parquet dataset.
}
\description{
Initiate a CBM4 spatial parquet dataset with study area geographic metadata.
}
