% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_write_inventory.R
\name{cbm4_write_inventory}
\alias{cbm4_write_inventory}
\title{CBM4 write inventory}
\usage{
cbm4_write_inventory(
  cbm4_data = NULL,
  grid_meta,
  grid_rast,
  cohortDT,
  classifiers,
  dataset_name = "inventory",
  dataset_path = file.path(cbm4_data, dataset_name),
  cbm_defaults_db = getOption("CBM4r.db.path"),
  ...
)
}
\arguments{
\item{cbm4_data}{character.
Path to CBM4 spatial parquet datasets directory.
May be omitted if full paths to datasets are provided.}

\item{grid_meta}{data.table. Grid metadata set with \code{cbm4_set_grid_meta}.}

\item{grid_rast}{terra \code{SpatRaster}. Grid defining the study area.}

\item{cohortDT}{data.table. Cohort inventory.}

\item{classifiers}{character.
Column names of inventory identifiers.}

\item{dataset_name}{character.
Name of the CBM4 spatial parquet dataset.}

\item{dataset_path}{character.
Path to the CBM4 spatial parquet dataset.
Defaults to \code{file.path(cbm4_data, dataset_name)}}

\item{cbm_defaults_db}{character.
Path to CBM defaults SQLite database.}

\item{...}{arguments to \code{\link{cbm4_format_inventory}}}
}
\value{
\code{NULL}. Data will be written to the CBM4 spatial parquet dataset.
}
\description{
Write inventory to a CBM4 spatial parquet dataset.
}
