% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_write_inventory.R
\name{cbm4_write_inventory}
\alias{cbm4_write_inventory}
\title{CBM4 write inventory}
\usage{
cbm4_write_inventory(
  cbm4_data = NULL,
  cbm_defaults_db,
  cohortDT,
  classifiers,
  grid_rast,
  grid_meta,
  dataset_name = "inventory",
  dataset_path = file.path(cbm4_data, dataset_name),
  ...
)
}
\arguments{
\item{cbm4_data}{character.
Path to CBM4 spatial parquet datasets directory.
May be omitted if full paths to datasets are provided.}

\item{cbm_defaults_db}{character.
Path to CBM defaults SQLite database.}

\item{cohortDT}{data.table. Cohort inventory.}

\item{classifiers}{character.
Column names of inventory identifiers.}

\item{grid_rast}{terra \code{SpatRaster}. Grid defining the study area.}

\item{grid_meta}{data.table. Grid metadata.
Required columns: \code{pixel_index},
\code{admin_boundary}, \code{admin_abbrev}, or \code{admin_boundary_id}.
\code{eco_boundary} or \code{eco_boundary_id},
\code{spatial_unit}.
Optional columns: \code{chunk_index}, \code{raster_index},
\code{afforestation_pre_type}, \code{historic_disturbance_type}, \code{last_pass_disturbance_type}.}

\item{dataset_name}{character.
Name of the CBM4 spatial parquet dataset.}

\item{dataset_path}{character.
Path to the CBM4 spatial parquet dataset.
Defaults to \code{file.path(cbm4_data, dataset_name)}}

\item{...}{arguments to \code{\link{cbm4_format_inventory}} or \code{\link{set_grid_meta}}}
}
\value{
\code{NULL}. Data will be written to the CBM4 spatial parquet dataset.
}
\description{
Write inventory to a CBM4 spatial parquet dataset.
}
