% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_write_simulation_inventory.R
\name{cbm4_write_simulation_inventory}
\alias{cbm4_write_simulation_inventory}
\title{CBM4 write simulation inventory}
\usage{
cbm4_write_simulation_inventory(
  cbm4_data = NULL,
  cohortDT,
  timestep,
  classifiers = NULL,
  template_name = "inventory",
  template_path = file.path(cbm4_data, template_name),
  dataset_name = "simulation",
  dataset_path = file.path(cbm4_data, dataset_name),
  ...
)
}
\arguments{
\item{cbm4_data}{character.
Path to CBM4 spatial parquet datasets directory.
May be omitted if full paths to datasets are provided.}

\item{cohortDT}{data.table. Cohort inventory.}

\item{timestep}{integer. Simulation timestep with 1 representing the first year.}

\item{classifiers}{character.
Column names of inventory identifiers.}

\item{template_name}{character.
Name of a CBM4 spatial parquet dataset to use as a template for the new dataset.}

\item{template_path}{character.
Path to CBM4 spatial parquet dataset to use as a template for the new dataset.
Defaults to \code{file.path(cbm4_data, template_name)}}

\item{dataset_name}{character.
Name of the CBM4 spatial parquet dataset.}

\item{dataset_path}{character.
Path to the CBM4 spatial parquet dataset.
Defaults to \code{file.path(cbm4_data, dataset_name)}}

\item{...}{arguments to \code{\link{cbm4_format_simulation_inventory}}}
}
\value{
\code{NULL}. Data will be written to the CBM4 spatial parquet dataset.
}
\description{
Write inventory to a simulation CBM4 spatial parquet dataset.
}
