% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_write_parameters.R
\name{cbm4_write_spinup_parameters}
\alias{cbm4_write_spinup_parameters}
\title{CBM4 write spinup parameters}
\usage{
cbm4_write_spinup_parameters(...)
}
\arguments{
\item{...}{arguments to \code{\link{cbm4_write_parameters}}}
}
\value{
\code{NULL}. Data will be written to the CBM4 spatial parquet dataset.
}
\description{
Write spinup parameters to a CBM4 spatial parquet dataset.
}
