% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_write_parameters.R
\name{cbm4_write_spinup_parameters}
\alias{cbm4_write_spinup_parameters}
\title{CBM4 write spinup parameters}
\usage{
cbm4_write_spinup_parameters(
  cbm4_data = NULL,
  cbm_defaults_db,
  grid_rast = NULL,
  grid_chunks = 1,
  template_name = NULL,
  template_path = file.path(cbm4_data, template_name),
  dataset_name = "spinup_parameters",
  dataset_path = file.path(cbm4_data, dataset_name),
  ...
)
}
\arguments{
\item{cbm4_data}{character.
Path to CBM4 spatial parquet datasets directory.
May be omitted if full paths to datasets are provided.}

\item{cbm_defaults_db}{character.
Path to CBM defaults SQLite database.}

\item{grid_rast}{terra \code{SpatRaster}. Grid defining the study area.}

\item{grid_chunks}{integer.
The number of chunks to partition the grid into for parallel processing.}

\item{template_name}{character.
Name of a CBM4 spatial parquet dataset to use as a template for the new dataset.}

\item{template_path}{character.
Path to CBM4 spatial parquet dataset to use as a template for the new dataset.
Defaults to \code{file.path(cbm4_data, template_name)}}

\item{dataset_name}{character.
Name of the CBM4 spatial parquet dataset.}

\item{dataset_path}{character.
Path to the CBM4 spatial parquet dataset.
Defaults to \code{file.path(cbm4_data, dataset_name)}}

\item{...}{arguments to \code{\link{cbm4_format_increments}}}
}
\value{
\code{NULL}. Data will be written to the CBM4 spatial parquet dataset.
}
\description{
Write spinup parameters to a CBM4 spatial parquet dataset.
}
