% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm4_write_parameters.R
\name{cbm4_write_step_parameters}
\alias{cbm4_write_step_parameters}
\title{CBM4 write step parameters}
\usage{
cbm4_write_step_parameters(...)
}
\arguments{
\item{...}{arguments to \code{\link{cbm4_write_parameters}}}
}
\value{
\code{NULL}. Data will be written to the CBM4 spatial parquet dataset.
}
\description{
Write step parameters to a CBM4 spatial parquet dataset.
}
